[exporter/loki] Remove deprecated configuration paramaters.
