+++ .chloggen/loki-remove-deprecated-parameters.yaml
@@ -1,0 +1,17 @@
+# One of 'breaking', 'deprecation', 'new_component', 'enhancement', 'bug_fix'
+change_type: breaking
+
+# The name of the component, or a single word describing the area of concern, (e.g. filelogreceiver)
+component: exporter/lokiexporter
+
+# A brief description of the change.  Surround your text with quotes ("") if it needs to start with a backtick (`).
+note: Remove deprecated Loki exporters parameters.
+
+# One or more tracking issues related to the change
+issues: [15653]
+
+# (Optional) One or more lines of additional information to render under the primary note.
+# These lines will be padded with 2 spaces and then inserted directly into the document.
+# Use pipe (|) for multiline entries.
+subtext: The support of `labels.{attributes/resource}`, `labels.record`, `tenant`,
+  | and `format` configuration parameters are dropped in favor of attribute hints way of configuring the exporter.
