+++ exporter/lokiexporter/config_test.go
@@ -88,79 +88,3 @@ func TestLoadConfigNewExporter(t *testing.T) {
 		})
 	}
 }
-
-func TestIsLegacy(t *testing.T) {
-	testCases := []struct {
-		desc    string
-		cfg     *Config
-		outcome bool
-	}{
-		{
-			// the default mode for an empty config is the new logic
-			desc: "not legacy",
-			cfg: &Config{
-				HTTPClientSettings: confighttp.HTTPClientSettings{
-					Endpoint: "https://loki.example.com",
-				},
-			},
-			outcome: false,
-		},
-		{
-			desc: "format is set to body",
-			cfg: &Config{
-				HTTPClientSettings: confighttp.HTTPClientSettings{
-					Endpoint: "https://loki.example.com",
-				},
-				Format: stringp("body"),
-			},
-			outcome: true,
-		},
-		{
-			desc: "a label is specified",
-			cfg: &Config{
-				HTTPClientSettings: confighttp.HTTPClientSettings{
-					Endpoint: "https://loki.example.com",
-				},
-				Labels: &LabelsConfig{
-					Attributes: map[string]string{"some_attribute": "some_value"},
-				},
-			},
-			outcome: true,
-		},
-		{
-			desc: "a tenant is specified",
-			cfg: &Config{
-				HTTPClientSettings: confighttp.HTTPClientSettings{
-					Endpoint: "https://loki.example.com",
-				},
-				Tenant: &Tenant{
-					Source: "static",
-					Value:  "acme",
-				},
-			},
-			outcome: true,
-		},
-		{
-			desc: "a tenant ID is specified",
-			cfg: &Config{
-				HTTPClientSettings: confighttp.HTTPClientSettings{
-					Endpoint: "https://loki.example.com",
-				},
-				TenantID: stringp("acme"),
-			},
-			outcome: true,
-		},
-	}
-	for _, tC := range testCases {
-		t.Run(tC.desc, func(t *testing.T) {
-			assert.Equal(t, tC.outcome, tC.cfg.isLegacy())
-
-			// all configs from this table test are valid:
-			assert.NoError(t, tC.cfg.Validate())
-		})
-	}
-}
-
-func stringp(str string) *string {
-	return &str
-}
