+++ exporter/lokiexporter/factory_test.go
@@ -1,0 +1,78 @@
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package lokiexporter
+
+import (
+	"context"
+	"fmt"
+	"net/http"
+	"testing"
+
+	"github.com/stretchr/testify/require"
+	"go.opentelemetry.io/collector/component/componenttest"
+	"go.opentelemetry.io/collector/config/confighttp"
+)
+
+const (
+	validEndpoint = "http://loki:3100/loki/api/v1/push"
+)
+
+func TestExporter_new(t *testing.T) {
+	t.Run("with valid config", func(t *testing.T) {
+		config := &Config{
+			HTTPClientSettings: confighttp.HTTPClientSettings{
+				Endpoint: validEndpoint,
+			},
+		}
+		exp := newExporter(config, componenttest.NewNopTelemetrySettings())
+		require.NotNil(t, exp)
+	})
+}
+
+func TestExporter_startReturnsNillWhenValidConfig(t *testing.T) {
+	config := &Config{
+		HTTPClientSettings: confighttp.HTTPClientSettings{
+			Endpoint: validEndpoint,
+		},
+	}
+	exp := newExporter(config, componenttest.NewNopTelemetrySettings())
+	require.NotNil(t, exp)
+	require.NoError(t, exp.start(context.Background(), componenttest.NewNopHost()))
+}
+
+func TestExporter_startReturnsErrorWhenInvalidHttpClientSettings(t *testing.T) {
+	config := &Config{
+		HTTPClientSettings: confighttp.HTTPClientSettings{
+			Endpoint: "",
+			CustomRoundTripper: func(next http.RoundTripper) (http.RoundTripper, error) {
+				return nil, fmt.Errorf("this causes HTTPClientSettings.ToClient() to error")
+			},
+		},
+	}
+	exp := newExporter(config, componenttest.NewNopTelemetrySettings())
+	require.NotNil(t, exp)
+	require.Error(t, exp.start(context.Background(), componenttest.NewNopHost()))
+}
+
+func TestExporter_stopAlwaysReturnsNil(t *testing.T) {
+	config := &Config{
+		HTTPClientSettings: confighttp.HTTPClientSettings{
+			Endpoint: validEndpoint,
+		},
+	}
+	exp := newExporter(config, componenttest.NewNopTelemetrySettings())
+	require.NotNil(t, exp)
+	require.NoError(t, exp.stop(context.Background()))
+}
