--- exporter/lokiexporter/testdata/config_legacy.yaml
@@ -1,47 +1,0 @@
-loki:
-  endpoint: "https://loki:3100/loki/api/v1/push"
-  tenant_id: "example"
-  labels:
-    attributes:
-      container.name: "container_name"
-      k8s.cluster.name: "k8s_cluster_name"
-      severity: "severity"
-loki/json:
-  endpoint: "https://loki:3100/loki/api/v1/push"
-  tenant_id: "example"
-  format: "json"
-  labels:
-    record:
-      traceID: traceid
-loki/allsettings:
-  endpoint: "https://loki:3100/loki/api/v1/push"
-  tenant_id: "example"
-  tls:
-    insecure: true
-    ca_file: /var/lib/mycert.pem
-    cert_file: certfile
-    key_file: keyfile
-  timeout: 10s
-  read_buffer_size: 123
-  write_buffer_size: 345
-  sending_queue:
-    enabled: true
-    num_consumers: 2
-    queue_size: 10
-  retry_on_failure:
-    enabled: true
-    initial_interval: 10s
-    max_interval: 60s
-    max_elapsed_time: 10m
-  headers:
-    "X-Custom-Header": "loki_rocks"
-  labels:
-    record:
-      traceID: traceid
-    attributes:
-      container.name: "container_name"
-      k8s.cluster.name: "k8s_cluster_name"
-      severity: "severity"
-    resource:
-      resource.name: "resource_name"
-      severity: "severity"
